Rename directory holding SIMD  parallel keccak implementation
