--- src/chaining_value/keccak/mod.rs
@@ -1,2 +1,0 @@
-mod consts;
-mod keccakx2;
