+++ src/keccak/consts.rs
