+++ src/keccak/mod.rs
@@ -1,0 +1,4 @@
+mod consts;
+
+#[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
+pub mod keccakx2;
